Create define_a_class_to_represent_a_bank_account_include_the_following_members.java
Define a class to represent a bank account. Include the following members:
Data Members:
a.Name of the Depositor
b.Account Number
c.Type of Account
d.Balance amount in the account
Data Methods:
a.To assign initial values
b.To deposit an amount
c.To withdraw an amount
d.To display name and balance.